pyproject.toml
[project]
name = "file-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "google-adk>=1.15.1",
    "httpx>=0.28.1",
    "litellm>=1.77.5",
    "pytest>=8.4.2",
]

[tool.ruff]
line-length = 100
indent-width = 2
target-version = "py313"

[tool.ruff.lint]
unfixable = []
per-file-ignores = {}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
